Adding XSLT files as a submodule.
This is to better manage the state of these files vis-a-vis
the tracking system.  Checking these in as submodules ensures we
have some way of tracking the state of these files as they change
over time.  Be aware that several of these files are stored in Fedora
as well, for on-the-fly transformations.  Always check the MD5 for
those objects' XSLT datastreams against the versions here.