Refactoring to eliminate phantom method calls.
Looks like some old code was calling a modele method that
has not existed for some time.  UnitsFilter is nowhere to be
found here or in Hydraulics.  Replacing with test that AFAIK
will duplicate the functionality described in the comments.

Also adding a virtual attribute 'mime-type' for MasterFiles and
subclasses, as that was being called deeper down.

All of this is to facilitate Hydra.desc(bibl) calls where we
are building MODS records based only on database information.