Hotfix: reading IndexDestination from Components
This is to accord with new practiced in Solr production, which
require a field 'released_facet' flagging a record for copying
into another index.  As Component Solr records are built via a
call to a Fedora disseminator, this value is passes parametrically
in the HTTP request.

TO DO: migration to ensure Components belong to an IndexDestination