Adding utility method to Unit class
This is in response to a third party developer who is
pulling data from our Fedora Repository.  Not knowing whether a
given MasterFile object (in Fedora) has a transcription
datastream, he was getting timeouts and 404 errors among
his requests.

This method allows me to quickly add whitespace boilerplate
transcriptions to our objects within a unit that lack such data,
and will update the corresponding objects if they exist in the repo.

Called from the console at the moment, but could easily be
added to an admin page.