Adding wrapper for restart scripts
Changes to views and other front-end behavior require only a touch
of the restart file (if you're using Passenger), but changes to
library code, models, etc., will require a restarting of all
ActiveMessenger pollers, which also use Tracksys.  This simple
Bash script is designed to simplify the 'restart all' operation.