Adding utility method for iView editing.
Creating new Components must happen from time to time, and I found
I needed to return to the old iView XML and add a <Set> to correspond
to this object.  This method allows a quick cut and paste from the new
(pidded) object, simplifying the XML edits.

Hopefully, in future this will be used as part of a more complex method
updating an iView XML file automatically, upon joinging a new Component
to a hierarchy.