Restoring broken JQuery animation on large MODS
This seems to have broken during a previous JQuery upgrade.
The JS additions aren't pretty, but there were context problems
passing the function reference through the DOM.  Wrapping the
function in an object (aka 'anobj') was an expedient solution.

Test on Component 500020, as it has a custom MODS file in it.