Fixes #151 (Transaction deadlocking on MySQL)
This has been tested during the creation of MasterFiles
for the Reed Yellow Fever Collection (Order 7001) and typically
recovered the transaction on the first retry.

Recommend you don't aimlessly click on TS pages looking for
messages, as this may increase the number of locks on tables
while the transaction is running.

Rule of thumb: units with >200 masterfiles should be left alone
for 10-20 minutes.